requirements.txt
# Django Core
Django>=4.1,<6.0
Pillow>=9.0
numpy>=1.21
joblib>=1.1

# Google OAuth Authentication
django-allauth>=0.57.0
requests>=2.28.0
PyJWT>=2.0.0
cryptography>=3.0
oauthlib>=3.0
requests-oauthlib>=1.0

# Environment variables
python-dotenv>=1.0.0

# Database (for production)
dj-database-url>=2.0.0

# ML dependencies
scikit-learn>=1.0
